+++ pkg/machine/qemu/command/command_test.go
@@ -1,3 +1,5 @@
+//go:build !darwin
+
 package command
 
 import (
