+++ pkg/machine/qemu/command/helpers.go
@@ -1,3 +1,5 @@
+//go:build !darwin
+
 package command
 
 import (
