+++ pkg/machine/qemu/command/qemu_command_test.go
@@ -1,4 +1,4 @@
-//go:build (amd64 && !windows) || (arm64 && !windows)
+//go:build ((amd64 && !windows) || (arm64 && !windows)) && !darwin
 
 package command
