+++ pkg/machine/qemu/machine_unsupported.go
@@ -1,3 +1,3 @@
-//go:build !amd64 && !arm64
+//go:build !amd64 && !arm64 && darwin
 
 package qemu
