+++ pkg/machine/qemu/options_freebsd_amd64.go
@@ -1,3 +1,5 @@
+//go:build freebsd && amd64
+
 package qemu
 
 var (
