+++ pkg/machine/qemu/options_freebsd_arm64.go
@@ -1,3 +1,5 @@
+//go:build freebsd && arm64
+
 package qemu
 
 var (
