+++ pkg/machine/qemu/options_linux_amd64.go
@@ -1,3 +1,5 @@
+//go:build linux && amd64
+
 package qemu
 
 var (
