+++ pkg/machine/qemu/options_linux_arm64.go
@@ -1,3 +1,5 @@
+//go:build linux && arm64
+
 package qemu
 
 import (
