+++ pkg/machine/qemu/options_windows_amd64.go
@@ -1,3 +1,5 @@
+//go:build windows && amd64
+
 package qemu
 
 var (
