+++ pkg/machine/qemu/options_windows_arm64.go
@@ -1,3 +1,5 @@
+//go:build windows && arm64
+
 package qemu
 
 var (
